Add resize events
Made-with: Cursor
